// src/config/configuration.enum.ts
export enum CONFIG {
  NODE_ENV = 'NODE_ENV',
  API_HOST = 'API_HOST',
  FRONTEND_URL = 'FRONTEND_URL',
  PORT = 'PORT',
  MONGODB_URI = 'MONGODB_URI',
  TOKEN_SECRET = 'TOKEN_SECRET',
  TOKEN_EXPIRES_IN = 'TOKEN_EXPIRES_IN',
  REFRESH_TOKEN_SECRET = 'REFRESH_TOKEN_SECRET',
  REFRESH_TOKEN_EXPIRES_IN = 'REFRESH_TOKEN_EXPIRES_IN',
  ENCRYPT_SECRET = 'ENCRYPT_SECRET',
  RESET_PASSWORD_TOKEN_EXPIRATION = 'RESET_PASSWORD_TOKEN_EXPIRATION',
  EMAIL_CONFIRMATION_TOKEN_EXPIRATION = 'EMAIL_CONFIRMATION_TOKEN_EXPIRATION',
  GOOGLE_CLIENT_ID = 'GOOGLE_CLIENT_ID',
  GOOGLE_CLIENT_SECRET = 'GOOGLE_CLIENT_SECRET',
  FACEBOOK_CLIENT_ID = 'FACEBOOK_CLIENT_ID',
  FACEBOOK_CLIENT_SECRET = 'FACEBOOK_CLIENT_SECRET',
  SMTP_HOST = 'SMTP_HOST',
  SMTP_PORT = 'SMTP_PORT',
  SMTP_USER = 'SMTP_USER',
  SMTP_SECRET = 'SMTP_SECRET',
  SMTP_FROM_EMAIL = 'SMTP_FROM_EMAIL',
  AWS_REGION = 'AWS_REGION',
  AWS_PUBLIC_BUCKET_NAME = 'AWS_PUBLIC_BUCKET_NAME',
  AWS_ACCESS_KEY_ID = 'AWS_ACCESS_KEY_ID',
  AWS_SECRET_ACCESS_KEY = 'AWS_SECRET_ACCESS_KEY',
  RD_STATION_CLIENT_ID = 'RD_STATION_CLIENT_ID',
  RD_STATION_CLIENT_SECRET = 'RD_STATION_CLIENT_SECRET',
  REDIS_HOST = 'REDIS_HOST',
  REDIS_PORT = 'REDIS_PORT',
  REDIS_PASSWORD = 'REDIS_PASSWORD',
  SEND_GRID_KEY = 'SEND_GRID_KEY',
  PROFILE_IMAGE_PATH = 'PROFILE_IMAGE_PATH',
  JOB_IMAGE_PATH = 'JOB_IMAGE_PATH',
}